Installer fix

+++ install.sh
@@ -1,12 +1,11 @@
 #!/data/data/com.termux/files/usr/bin/bash
 set -e
 
+echo "[+] Checking files..."
+[ -f bcbs ] || { echo "[ERROR] bcbs not found"; exit 1; }
+[ -f terminal.py ] || { echo "[ERROR] terminal.py not foun"; exit 1; }
+
 echo "[+] Instaling BCBS-NetHunter..."
-
-chmod +x bcbs
-chmod +x terminal.py
-
+chmod +x bcbs terminal.py
 cp bcbs $PREFIX/bin/bcbslauncher
-
-echo "[ OK ] BCBS NetHunter is ready to use"
-echo "To start type: bcbs"
+echo "[ OK ] Type: bcbslauncher"
